// client/src/lib/types.ts
export type Card = {
  id: number;
  cardId: string;
  imageUrl: string;
  creatorId: number;
  creatorName: string;
  creatorSequence: number;
  createdAt: string;
  timesPlayed: number;
  timesCollected: number;
  timesDiscovered: number;
  discoveredAt?: string;
  collectedAt?: string;
};

export type Deck = {
  id: number;
  ownerId: number;
  ownerName: string;
  name: string;
  description: string;
  createdAt: string;
  timesCollected: number;
  cards: Card[];
};
// ...
export type RoomPlayer = {
  id: "A" | "B" | "C" | "D";
  userId: number;
  deckId: number;
};

export type GameSubmission = {
  playerId: RoomPlayer["id"];
  card: Card;
  isStoryCard: boolean;
};

export type GameVote = {
  voterId: RoomPlayer["id"];
  submissionIndex: number;
};

export type GameRound = {
  storytellerId: RoomPlayer["id"];
  clue: string;
  submissions: GameSubmission[];
  anonymousCards: GameSubmission[];
  votes: GameVote[];
  revealed: boolean;
  discoveredCards: Card[];
};